route, node: create stub files for Darwin builds
Running unit tests on Mac OS environments for many packages in Cilium has not
been possible due to the fact that there are dependencies which depend on
variables that are defined only for Linux architectures, e.g. netlink.FAMILY_V6.

Refactor the node and route packages to have Linux-specific and Darwin-specific
files, by using build constraints. For Darwin builds, stubs are created, while
the implementation used by the cilium-agent is factored out into files built
for non-Darwin builds. This allows for running of unit tests on the Mac for
`pkg/policy`, for instance.

No functional change is intended by this.

Signed-off by: Ian Vernon <ian@cilium.io>
